Rework Makefile
Makefile is now much cleaner. Also added beginnings of AST printing.
Still doesn't work properly though. Made some minor changes throughout
the code to quash warnings being thrown due to new compiler flags. Added
initial declaration to test2.txt so it's semantically correct now. Was
passing semantic analysis before that anyway, so something's wrong
clearly.

+++ src/symbol.c
@@ -16,7 +16,7 @@ static unsigned int stringHash(const char *string, int size)
 {
 	unsigned int h = 0;
 	
-	for (int i = 0; i < strlen(string); i++)
+	for (int i = 0; (size_t)i < strlen(string); i++)
 		h = (31 * h + string[i]) % size;
 
 	return h;
